requirements.txt
opencv-python
zipfile36
Pillow
mediapipe